api/requirements.txt
fastapi
uvicorn
websocket_client
requests
requests_toolbelt
openai
python-docx
python-multipart